paper-finder/requirements.txt
# Paper Finder Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
requests>=2.28.0
pandas>=1.5.0
openpyxl>=3.0.0
PyYAML>=6.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Optional dependencies (uncomment if needed)
# tqdm>=4.64.0      # Progress bars
# colorama>=0.4.0   # Colored output

# Development dependencies (uncomment for development)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=5.0.0
